+++ etc/spack/defaults/config.yaml
@@ -21,11 +21,11 @@ config:
     projections:
       all: "${ARCHITECTURE}/${COMPILERNAME}-${COMPILERVER}/${PACKAGE}-${VERSION}-${HASH}"
     # install_tree can include an optional padded length (int or boolean)
-    # default is False
+    # default is False (do not pad)
     # if padded_length is True, Spack will pad as close to the system max path
     # length as possible
     # if padded_length is an integer, Spack will pad to that many characters,
-    # assuming it his higher than the length of the install_tree root.
+    # assuming it is higher than the length of the install_tree root.
     # padded_length: 128
